integrity-checker: Pull out write set comparison functions into a class.
And move all the tests to its test class.
