+++ ledger/participant-state/kvutils/tools/src/main/scala/ledger/participant/state/kvutils/tools/integritycheck/CommitStrategySupport.scala
@@ -4,7 +4,6 @@
 package com.daml.ledger.participant.state.kvutils.tools.integritycheck
 
 import akka.stream.Materializer
-import com.daml.ledger.participant.state.kvutils.Raw
 import com.daml.ledger.participant.state.kvutils.export.{SubmissionInfo, WriteSet}
 import com.daml.ledger.participant.state.v1.ReadService
 import com.daml.ledger.validator.StateKeySerializationStrategy
@@ -31,21 +30,4 @@ trait CommitStrategySupport[LogResult] {
   )(implicit materializer: Materializer): Future[WriteSet]
 
   def newReadServiceFactory(): ReplayingReadServiceFactory
-
-  /** Determines if there's an actual difference and tries to explain it in case there is.
-    * A None return value signals that no difference should be signaled to the user.
-    */
-  def explainMismatchingValue(
-      key: Raw.Key,
-      expectedValue: Raw.Value,
-      actualValue: Raw.Value,
-  ): Option[String]
-
-  /** Validates that a single write set entry is readable.
-    *
-    * @param rawKey   The serialized key.
-    * @param rawValue The serialized value.
-    * @return `Right(())` if the entry is valid, or `Left(message)` with an explanation otherwise.
-    */
-  def checkEntryIsReadable(rawKey: Raw.Key, rawValue: Raw.Value): Either[String, Unit]
 }
