+++ ledger/participant-state/kvutils/tools/src/main/scala/ledger/participant/state/kvutils/tools/integritycheck/IntegrityChecker.scala
@@ -10,12 +10,9 @@ import akka.stream.Materializer
 import akka.stream.scaladsl.{Sink, Source}
 import com.codahale.metrics.{ConsoleReporter, MetricRegistry}
 import com.daml.dec.DirectExecutionContext
-import com.daml.ledger.participant.state.kvutils
-import com.daml.ledger.participant.state.kvutils.Raw
 import com.daml.ledger.participant.state.kvutils.export.{
   LedgerDataImporter,
   ProtobufBasedLedgerDataImporter,
-  WriteSet,
 }
 import com.daml.ledger.participant.state.v1.{ParticipantId, ReadService}
 import com.daml.ledger.resources.{ResourceContext, ResourceOwner}
@@ -26,7 +23,6 @@ import com.daml.platform.configuration.ServerRole
 import com.daml.platform.indexer.{IndexerConfig, IndexerStartupMode, JdbcIndexer}
 import com.daml.platform.store.dao.events.LfValueTranslation
 
-import scala.PartialFunction.condOpt
 import scala.concurrent.duration.Duration
 import scala.concurrent.{ExecutionContext, ExecutionContextExecutorService, Future}
 import scala.util.{Failure, Success}
@@ -37,6 +33,8 @@ class IntegrityChecker[LogResult](
   private val metricRegistry = new MetricRegistry
   private val metrics = new Metrics(metricRegistry)
   private val commitStrategySupport = commitStrategySupportBuilder(metrics)
+  private val writeSetComparison =
+    new WriteSetComparison(commitStrategySupport.stateKeySerializationStrategy)
 
   import IntegrityChecker._
 
@@ -184,7 +182,7 @@ class IntegrityChecker[LogResult](
             + s" writeSetSize=${expectedWriteSet.size}"
         )
         expectedWriteSet.foreach { case (key, value) =>
-          val result = commitStrategySupport.checkEntryIsReadable(key, value)
+          val result = writeSetComparison.checkEntryIsReadable(key, value)
           result.left.foreach { message =>
             throw new UnreadableWriteSetException(message)
           }
@@ -199,9 +197,13 @@ class IntegrityChecker[LogResult](
                 actualWriteSet
             actualReadServiceFactory.appendBlock(orderedActualWriteSet)
 
-            if (config.performByteComparison) {
-              compareWriteSets(expectedWriteSet, orderedActualWriteSet)
-            }
+            if (config.performByteComparison)
+              writeSetComparison.compareWriteSets(expectedWriteSet, orderedActualWriteSet) match {
+                case None =>
+                  println("OK".green)
+                case Some(message) =>
+                  throw new ComparisonFailureException(message)
+              }
           }
         } else {
           Future.unit
@@ -213,94 +215,6 @@ class IntegrityChecker[LogResult](
         println()
       }
   }
-
-  private def compareWriteSets(expectedWriteSet: WriteSet, actualWriteSet: WriteSet): Unit =
-    if (expectedWriteSet == actualWriteSet) {
-      println("OK".green)
-    } else {
-      val messageMaybe =
-        if (expectedWriteSet.size == actualWriteSet.size) {
-          compareSameSizeWriteSets(expectedWriteSet, actualWriteSet)
-        } else {
-          Some(
-            Seq(
-              Seq(
-                s"Expected write-set of size ${expectedWriteSet.size} vs. ${actualWriteSet.size}."
-              ),
-              Seq("Expected:"),
-              writeSetToStrings(expectedWriteSet),
-              Seq("Actual:"),
-              writeSetToStrings(actualWriteSet),
-            ).flatten.mkString(System.lineSeparator())
-          )
-        }
-      messageMaybe.foreach { message =>
-        throw new ComparisonFailureException(message)
-      }
-    }
-
-  private def writeSetToStrings(writeSet: WriteSet): Seq[String] =
-    writeSet.view.map((writeItemToString _).tupled).toVector
-
-  private def writeItemToString(key: Raw.Key, value: Raw.Value): String = {
-    val keyString = commitStrategySupport.stateKeySerializationStrategy.deserializeStateKey(key)
-    val valueString = kvutils.Envelope.open(value)
-    s"$keyString -> $valueString"
-  }
-
-  private[tools] def compareSameSizeWriteSets(
-      expectedWriteSet: WriteSet,
-      actualWriteSet: WriteSet,
-  ): Option[String] = {
-    val differencesExplained = expectedWriteSet
-      .zip(actualWriteSet)
-      .map { case ((expectedKey, expectedValue), (actualKey, actualValue)) =>
-        if (expectedKey == actualKey && expectedValue != actualValue) {
-          explainDifference(expectedKey, expectedValue, actualValue).map { explainedDifference =>
-            Seq(
-              s"expected value:    ${rawHexString(expectedValue)}",
-              s" vs. actual value: ${rawHexString(actualValue)}",
-              explainedDifference,
-            )
-          }
-        } else if (expectedKey != actualKey) {
-          Some(
-            Seq(
-              s"expected key:    ${rawHexString(expectedKey)}",
-              s" vs. actual key: ${rawHexString(actualKey)}",
-            )
-          )
-        } else {
-          None
-        }
-      }
-      .map(_.toList)
-      .filterNot(_.isEmpty)
-      .flatten
-      .flatten
-      .mkString(System.lineSeparator())
-    condOpt(differencesExplained.isEmpty) { case false =>
-      differencesExplained
-    }
-  }
-
-  private def explainDifference(
-      key: Raw.Key,
-      expectedValue: Raw.Value,
-      actualValue: Raw.Value,
-  ): Option[String] =
-    kvutils.Envelope
-      .openStateValue(expectedValue)
-      .toOption
-      .map { expectedStateValue =>
-        val stateKey =
-          commitStrategySupport.stateKeySerializationStrategy.deserializeStateKey(key)
-        val actualStateValue = kvutils.Envelope.openStateValue(actualValue)
-        s"""|State key: $stateKey
-            |Expected: $expectedStateValue
-            |Actual: $actualStateValue""".stripMargin
-      }
-      .orElse(commitStrategySupport.explainMismatchingValue(key, expectedValue, actualValue))
 
   private def reportDetailedMetrics(metricRegistry: MetricRegistry): Unit = {
     val reporter = ConsoleReporter
@@ -343,9 +257,6 @@ class IntegrityChecker[LogResult](
 object IntegrityChecker {
   type CommitStrategySupportFactory[LogResult] =
     (Metrics, ExecutionContext) => CommitStrategySupport[LogResult]
-
-  def rawHexString(raw: Raw.Bytes): String =
-    raw.bytes.toByteArray.map(byte => "%02x".format(byte)).mkString
 
   abstract class CheckFailedException(message: String) extends RuntimeException(message)
